restructure project to have a package hanabi_ai under root
This appears to solve the import problem, as well as
cleans up some issues I was having with the project
