+++ hanabi_ai/model/hanabi_table.py
