+++ sample.sh
@@ -1,5 +1,5 @@
 python playgame.py "single" \
-    "ai.example_discarder.Discarder" \
-    "ai.example_discarder.Discarder" \
-    --log_dir "results/results.txt" \
-    --log_stderr "results/errors.txt"
+    "hanabi_ai.players.example_discarder.Discarder" \
+    "hanabi_ai.players.example_discarder.Discarder" \
+    --log_dir "game_logs/results.txt" \
+    --log_stderr "game_logs/errors.txt"
