+++ sample_tournament.sh
@@ -1,6 +1,6 @@
 python playgame.py "tournament" \
-    "ai.example_discarder.Discarder" \
-    "ai.example_discarder.Discarder" \
-    "ai.example_discarder.Discarder" \
-    --log_dir "results/tournament_results.txt" \
-    --log_stderr "results/tournament_errors.txt"
+    "hanabi_ai.players.example_discarder.Discarder" \
+    "hanabi_ai.players.example_discarder.Discarder" \
+    "hanabi_ai.players.example_discarder.Discarder" \
+    --log_dir "game_logs/tournament_results.txt" \
+    --log_stderr "game_logs/tournament_errors.txt"
